refactor: extract reusable MCP composition compiler
Ref: ED-24967
